pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kim-convergence"
version = "0.1.0"
requires-python = ">=3.7"
dependencies = [
  "numpy",
  "omegaconf",
  "hydra-core"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
